requirements.txt
streamlit==1.35.0
gspread==5.12.4
oauth2client==4.1.3
pandas==2.2.2
beautifulsoup4==4.12.3
requests==2.32.2
selenium==4.20.0
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
